move cct to its own folder
