+++ src/DevMenu/ui/CodingContractsDev.tsx
@@ -9,8 +9,8 @@ import Typography from "@mui/material/Typography";
 import Button from "@mui/material/Button";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
 import MenuItem from "@mui/material/MenuItem";
-import { generateContract, generateRandomContract, generateRandomContractOnHome } from "../../CodingContractGenerator";
-import { isCodingContractName } from "../../data/codingcontracttypes";
+import { generateContract, generateRandomContract, generateRandomContractOnHome } from "../../CodingContract/ContractGenerator";
+import { isCodingContractName } from "../../CodingContract/ContractTypes";
 import { CodingContractName } from "@enums";
 
 export function CodingContractsDev(): React.ReactElement {
